Corrected the location of the update table part
In my previous PR I put the documentation to update a table in the wrong spot - This PR fixes that. It's now where it should belong.
Also, on the same part, I fixed the code where a variable was silently overwritten. It was working, however, it's a bad practice and could lead to unexpected behaviour (Trying to access `$data` under those lines).